+++ docs/history.rst
@@ -1,5 +1,6 @@
 0.3
   * bugfix: pip installation issue :ghissue:`8`.
+  * new: TelnetServer.connected property is a Future set when first connected.
 
 0.2
   * enhancement: various example programs were included in this release.